avformat/matroskadec: Use av_fast_realloc_array()
instead of av_fast_realloc() for allocating an array. It has the
advantage of doing it's own overflow checks and does not overallocate
unnecessarily: It allocates exactly one element if one element is
desired. This is advantageous for CueTrackPositions: While it is
allowed (and supported) to have more than one of them in a CuePoint,
there is typically only one of them, so that overallocating is a waste.
The memory used for a file with 540000 cues entries therefore decreased
from 69624 KB to 52592 KB.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>