avformat/matroskaenc: Remove special code for writing subtitles
Once upon a time, mkv_write_block() did what its name says: Only write a
(Simple)Block, not a BlockGroup which is needed for subtitles to convey
the duration. But with the introduction of support for writing
BlockAdditions and DiscardPadding (both of which require a BlockGroup),
mkv_write_block() can also open and close a BlockGroup on its own. This
naturally led to some code duplication which is removed in this commit.

This new code leads to one regression: It always uses eight bytes for
the BlockGroup's length field, whereas the earlier code usually used the
lowest amount of bytes needed. This will be fixed in a future commit.

This temporary regression is also the reason for a change to the
binsub-mksenc fate test.

The reason for putting the code for opening and closing BlockGroups in
mkv_write_block() rather than factoring it out are that writing the
least amount of bytes requires a knowledge of the size of the data, i.e.
it needs to come after the various conversions (e.g. from annex b H.264
to mp4-style annex b) have taken place.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>