avformat/matroskaenc: Support attached pics not initially available
This commit makes the Matroska muxer write streams with attached_pic
disposition (other than timed thumbnails) as AttachedFile if the output
is a seekable, ordinary file. If there are attached pics missing when
writing the header, then the Attachments will be written when writing
the trailer at the end of the file (and referenced in the SeekHead).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>