avformat/hevc: Clarify documentation of filter_ps
and use it to avoid unnecessary initializations and frees in
ff_hevc_annexb2mp4() and ff_hevc_annexb2mp4_buf(): Even if the caller
does not want to strip HEVC parameter sets away when converting from
Annex B to mp4, a pointer has been initialized to NULL and later freed
while still being NULL despite having not been used at all.

The reason for this is that there is still a jump to the end of the
function in order to set a variable (returned via a pointer argument)
containing the number of parameter set NAL units stripped away to zero.
But if the caller does not want to strip these parameter sets away, it
is pointless to set this variable. Document this behaviour and remove
the jump.

Also document that said number won't be set on error (in case of
ff_hevc_annexb2mp4_buf(), this was actually the behaviour anyway) and
remove the jump to the end on error, too; this also allows to avoid
initializing the pointer mentioned above.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>