avformat/aviobuf: Add function to reserve size in a dynamic buffer
This is useful to avoid later reallocations if one has a good guess
of the size that will eventually be written.

The type of the parameter used for the size to reserve is size_t despite
the dynamic buffer API being based on integers. This is done because it
is a size, in order to avoid casts in ffio_dyn_buf_reserve() as well as
for potential future extensions of the dynamic buffer API. But
allocations beyond INT_MAX are of course not allowed.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>