avformat/avc, hevc: Write data directly into dynamic buffer
When an AVIOContext is not in direct mode, avio_write() first copies all
the data into the write buffer; when this buffer is full, its contents
are written. And the buffer is filled again. This also happens for
dynamic buffers and it is costly if one mostly wants to perform big
writeouts. This is the case when one converts Annex B AVC/HEVC data to
mp4, so it makes sense to set this flag.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>