avformat/flacenc: Use attached pics given via st->attached_pic
Using these attached pics has the potential advantage of reducing the
delay of writing the attached pics (if the user already supplied them).

Furthermore, it also eliminate a memleak that is caused by the stream's
priv_data (which was used to store the attached pics before this commit)
simply being freed as an opaque buffer and not freed as an AVPacket.
(In contrast, AVStream.attached_pic is of course freed as a packet.)
This memleak could appear if the trailer was not written or if writing
an attached pic returned an error and the AV_EF_EXPLODE error
recognition flag was set.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>