avformat/segment: Propagate attached pic
This commit makes it possible to send attached pics so early to the
muxer that they are available when writing the header. This is
especially important when the header is written to a file of its own
(with the segment_header_filename option), as the usual method of
providing the attached pic (namely sending it as one of the first packets
to be muxed) fails in this scenario.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>