avformat/matroskaenc: Write Tags and Attachments early when possible
If some attached pictures are missing when writing the header,
the Attachments element will be written at the end, after the Clusters
when writing the trailer. Yet given that the Cluster is not written
immediately, but into a buffer in memory before being output all at
once, one can still write the Attachments element before the Clusters if
one receives all outstanding attached pictures before outputting the
first Cluster. This commit implements this.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>