avfilter/af_channelmap: Fix double-free of AVFilterChannelLayouts on error
The query_formats function of the channelmap filter tries to allocate a list
of channel layouts which on success are attached to more permanent objects
(an AVFilterLink) for storage afterwards. If attaching succeeds, the link
becomes one of the common owners (in this case, the only owner) of the list.
Yet if the list has been successfully attached to teh link and an error
happens lateron, the list was manually freed, which is wrong, because
it is owned by its link so that the link's pointer to the list will become
dangling and there will be a double-free/use-after-free when the link
is later cleaned up automatically.

This commit fixes this by removing the custom freeing code; this will
temporarily add a leaking codepath (if attaching the list fails, the list
will leak), but this will be fixed soon by making sure that an
AVFilterChannelLayouts without owner will be automatically freed when
attaching it to an AVFilterLink fails.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>