avfilter/af_aiir: Avoid temporary buffer when drawing image
When creating the output video, a temporary buffer is used whose elments
are only used three times: Once to store a value in it, once more to add
the value of the preceding element of the temporary buffer to the current
element and once more to add the current element #i to the element #i of
another bufffer; the two latter operations were performed in one loop,
whereas the first was performed in a loop of its own.

Yet these loops can be combined and the temporary buffer avoided. All
one needs to do is store the running total of the elements that
currently are in the temporary buffer and keep the preceding value of
the destination buffer (it is needed to calculate the next element of
the current temporary buffer, i.e. it is needed to know by how much to
increment the running total).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>