avfilter/af_headphone: Fix convolution
When the headphone filter performs its processing in the time domain,
it ought to perform a simple convolution. Given that the input is stored
in a cyclic buffer, it first has to check whether the desired content is
available in one continuous chunk or not. In the latter case it should
use all the samples of the ringbuffer until it reaches the wraparound
point and then use enough samples from the beginning of the ringbuffer
until it has enough samples.

Yet that is not what happened:

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>