avcodec/mpeg2_metadata_bsf: Set frame_pred_frame_dct when possible
When a frame's picture coding extension has frame_pred_frame_dct set,
every macroblock of said frame uses frame-based prediction and dct, so
that the relevant fields don't need to be coded in the bitstream on a
per-macroblock basis. Yet some encoders produce frames that are
compatible with frame_pred_frame_dct set, but don't set it, thereby
wasting bits. This commit changes this by adding an option to inspect
the frames and set the flag and modify the bitstream if possible.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>