avcodec/vc1: Use symbols table to decode AC coefficients
The VC1 decoder currently follows the spec very closely when decoding
AC coefficients: The spec uses tables with contiguous indexes. The last
of each of these is an escape code; the other codes encode three other
variables via tables: run, length and whether this is the last
coefficient.

This commit changes this: Given that both run and length fit into 6bits
each, all three of these fit into a VLC symbol (without touching the
sign bit which is needed because some of the VLCs used here aren't
complete).

This already makes the tables for the conversion from index to run etc.
superfluous; furthermore while just at it, this commit also switches to
ff_init_vlc_from_lengths(), thereby making the codes tables superfluous,
too (and said code tables were interleaved with the length tables, so
that the latter were using an uint32_t just as the former, wasting lots
of space). Moreover, the number of codes for each table differed and the
old approach padded the shorter ones to the length of the longer ones;
this is no longer done either. All in all, this saves about 11KB of
tables.

Finally, the symbol and length tables have been added to vc1data.c;
the interleaved codes and lengths tables which they replace was directly
in vc1.c since 3c715383ea7012ac69507e6b9189c98675c77461 in order to
make the VC1 parser compilable without decoder, but the data for the
other VLCs which are also only used in the decoder is in vc1data.c, so
moving it improves consistency.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>