avcodec/vc1: Use a symbols table to decode interlaced MV vector VLCs
The VC1 decoder currently follows the spec very closely when decoding
interlaced MV vector variable-length codes: The spec uses tables with
contiguous indexes. The last of these is an escape code, for every other
two values are encoded via (index + 1) / 9 and (index + 1) % 9. The
decoder did exactly this. This necessitated a branch to find out what
the current escape code was as well as the actual division/modulo.

This commit changes this by using an explicit symbols table for the
VLCs. The symbols are chosen so that the escape code always has the
value 0 (obviating the branch) and so that for every other code the
low nibble of the corresponding symbol contains (index + 1) % 9 whereas
the high nibble contains (index + 1) / 9.

Furthermore, the VLCs are now initialized with ff_init_vlc_from_lengths();
this allows to replace the uint32_t tables of codes by uint8_t tables of
symbols, thereby saving space on top of all the other improvements.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>