avcodec/vc1: Use symbols table to decode MV differentials
The VC1 decoder currently follows the spec very closely when decoding
MV differentials: The spec uses tables with contiguous indexes and
derives four values from it using a pseudo-code procedure: Two flags
and two values with a range of 0..5 each; the latter are obtained as the
result resp. the remainder of dividing the index by 6 in the spec and
in the earlier code.

Yet two flags can be encoded on two bits and the two values can be
encoded on three bits each. Therefore the new symbols uses the two
lowest bits for the flags; the two other values are coded separately
on three bits each. There is just one more minor complication:
The two three-bits values don't always exist; there is another mode.
Instead of using an arbitrary six bit value for the higher bits that
is different from all the values used by the ordinary mode this commit
instead uses a negative number for it; this is possible by shifting the
symbols as stored in the table (to make all fit into an uint8_t) and
then shifting the codes downwards when creating the table. The latter
is possible thanks to ff_init_vlc_from_lengths().

Switching to ff_init_vlc_from_lengths() also entails replacing 16bit
codes by 8bit symbols, thereby saving space.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>