avcodec/vc1: Use symbols table to parse macroblock mode
Expressions like array[get_vlc2()] can be optimized by using a symbols
table if the array is always the same for a given VLC. This requirement
is fulfilled for the VLCs used to parse the macroblock mode (MBMODE)
element for frame P and B pictures. MBMODE consists of four pieces of
information which can easily be encoded on one byte. Doing so not only
avoids a dereference, but also some space as the table containing this
information becomes redundant; furthermore, some of the VLCs used were
initialized with code tables of type uint16_t which are now replaced by
symbol tables of type uint8_t (thanks to ff_init_vlc_from_lengths()),
thereby saving space.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>