avcodec/mpeg4audio: Unavpriv and deduplicate mpeg4audio_sample_rates
avpriv_mpeg4audio_sample_rates has a size of 64B and it is currently
avpriv; a clone of it exists in aacenctab.h and from there it is inlined
in aacenc.c (which also uses the avpriv version) and in the FLV muxer.
This means that despite it being avpriv both libavformat as well as
libavcodec have copies already.

This situation is clearly suboptimal. Given the overhead of exporting
symbols (for x64 Elf/Linux/GNU: 2x2B version, 2x24B .dynsym, 24B .rela.dyn,
8B .got + twice the size of the name (here 31B)) the object is
unavprived, i.e. duplicated into libavformat when creating a shared
build; but the duplicates in the AAC encoder and FLV muxer are removed.

In order to simplify the changes to the Makefile a new CONFIG_EXTRA
variable mpeg4audio has been added; it is for the mpeg4audio_get_config
stuff and not for the mpeg4audio_sample_rates table.
Moving said table out of mpeg4audio.c also allowed to break some
dependencies (e.g. both the AAC encoder as well as the Matroska demuxer
actually don't need the mpeg4audio_get_config stuff).

Adding a new CONFIG_EXTRA variable requires rerunning configure;
in order to force this, some superfluous headers from
libavdevice/alldevices.c have been removed.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>