avfilter/formats: Avoid redundant counter
The ff_set_common_(formats|channel_layouts|samplerates) have to free
their list in case it doesn't have an owner; therefore they tracked
whether they attached it to an owner. But the list's refcount already
contains such a counter, so we don't have to keep track of whether we
have attached the list to an owner.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>