avfilter/qsvvpp: Fix double-free of AVFrame on error
An AVFilterPad's filter_frame callback in general and ff_filter_frame()
in particular take ownership of the provided frame and are responsible
for freeing it on error. Therefore it is wrong for
ff_qsvvpp_filter_frame() to free the frame itself on error again, as
all filter_frame callbacks in use are derived from ff_filter_frame().

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>