avfilter/internal: Add flag to free filtered AVFrames generically
Up until now, an AVFilterPad's filter_frame callback was always
responsible for passing the frame to the next filter or freeing it.
Yet there are several filter_frame callbacks that need to free this
frame on every exit, so it makes sense to do this generically.
This commit adds a flag which if set shifts the responsibility to free
the frame to the caller of filter_frame.

In the future this might be extended to avoid allocating+freeing the
frame for filters with this flag set.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>