avformat/aptxdec: Use ff_pcm_read_packet
It automatically honours block_align and it clears the
AV_PKT_FLAG_CORRUPT flag which up until now was always set
for the last packet if the number of input blocks was not
divisible by 256 (or equivalently: if the number of input
samples was not divisible by 1024).

It reads more data at once; this necessitated updates to
FATE reference files.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>