avfilter/avfilter: Remove redundant parsing code
avfilter_init_str() (via process_options()) both applies options
extracted from the given string directly to the relevant (private)
context as well as to an AVDictionary that is later given to
avfilter_init_dict() which applies these options again. This is
unnecessary, so leave applying the options to avfilter_init_dict();
this also has the advantage that it will be possible to report
all unrecognized options before erroring out in case of unrecognized
options, whereas the current code automatically stops after the first
such option.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>