avcodec/decode: Always claim the subtitle packet to be fully consumed
avcodec_decode_subtitle2() works like is already removed
audio/video counterparts: They return the number of bytes of
input consumed in a particular call and expect the user to
provide the unprocessed part of the current packet on the next
call like this:
while (1) {
    ret = avcodec_decode_subtitle2(avctx, sub, &got_sub_ptr, avpkt);
    if (ret < 0 || !got_sub_ptr)
        break;
    /* do something with sub */
    if (ret >= avpkt->size)
        break;
    pkt->data += ret; pkt->size -= ret;
}
Yet this never worked: The ccaption decoder simply returns 0 on success
and if recoding is used, the actually returned size is the size (in
bytes) of the recoded UTF-8 data; this might be less than the original
buffer size in which case the above algorithm would feed the decoder
again with a part of the original data (and the offsetted data does
not need to be valid according to the specified encoding).

Furthermore, the only subtitle decoder which not always consumes the
full packet on success is the pgs decoder which behaves a bit like
a parser in that it is possible to send it a packet and a part of the
next packet. This seems to be unintentional anyway. So this change
should not adversely affect this decoder.

Notice that both ffmpeg.c as well as demux.c already only call
avcodec_decode_subtitle2() once per packet.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>