avcodec/packet_internal: Move packet list API to libavformat
Libavcodec doesn't use this API at all and given that it now
no longer relies on any internals of AVPacket it can be moved
to its main user, libavformat. In case it is also used by
libavdevice, the small functions used by it are duplicated
into libavdevice in case of shared builds.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>