avcodec/get_bits: Don't let number of bits_left become invalid
When using the checked cached bitstream reader, it can currently
happen that the GetBitContext's number of bits_left becomes wrong,
namely off by UINT_MAX due to wraparound (if it were an int, it
would become negative).

The reason for this is that reading new input is performed through
show_bits() which does not update get_vlc2() if one is already
past the end of input*, so that the following skip_remaining() may
remove more bits than are actually present (it seems to have been
designed as if it were only to be called if there are enough bits
left). (The issue could be reproduced with
fraps-v5-bouncing-balls-partial.avi from the FATE suite if fraps
were modified to use a checked bitstream reader.)

At the moment, this does not seem to cause any issues:
The only place where bits_left is used as in the right
operand of a shift is when refilling and this doesn't
happen if one is already at the end. Furthermore, get_bits_count()
and get_bits_left() return proper values, because they return
an int and so being off by UINT_MAX doesn't matter.
Some functions that check whether the cache has enough bits left
will directly read from the cache.

This commit nevertheless intends to fix this by adding a variant
of refill_32() that always adds 32 bits to bits_left, but reports
if it has added fake bits (not read from the bitstream),
so that they can be compensated for. To do so, the first thing
get_vlc2() does is ensuring that at least 32 bits are in the cache
(even if fake bits), so that all the following operations
can be performed on the cache (the earlier code potentially
refilled the cache at every reading stage).

This is also beneficial for code size. In the following,
Clang means Clang 14 and GCC GCC 11.2:

           | Clang old | Clang new | GCC old | GCC new
_________________________________________________________
fraps      |    f84    |    f04    |   c9e   |   b7e
magicyuv   |   1f2f    |   1ecf    |  1b1b   |  1a49
mvha       |    c8a    |    bfa    |  11f4   |  1168
photocd    |   12d5    |   12f9    |  15ca   |  15aa
sheervideo |  14f2a    |  11778    | 103d3   | 10339
utvideodec |   35a1    |   34d1    |  3037   |  2ea7

(None of the above files (which are all files that use
the cached bitstream reader) call get_vlc2() with a max depth
of one; in this case, the refill_32() saved wins size-wise.
This is unfortunately not true in case of max_depth == 1,
where the added check leads to a slight codesize increase.)

*: The unchecked bitstream reader does not have this check
and is therefore unaffected from this issue.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>