avcodec/get_bits: Add ALIGNED32_LE_BITSTREAM_READER
Currently, some decoders swap the alignment of their input
buffer before reading it via an ordinary big-endian bitreader;
they swap it as 32bit words.*
Yet when using the cached bitstream reader, the swapping can be
done upon reading: Said bitstream reader already mostly reads
in wordlengths of 32 bit and reading these words as little-endian
instead of big-endian is equivalent to swapping the endianness
of the input buffer, i.e. makes swapping the endianness superfluous.

(This is not finished: skipping needs adjustments to honour alignment.)

*: Some decoders are based upon 16bit words for swapping.
This could also be done when reading, but only if one read
via 16 bit words.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>